+++ src/pages/game/helpers/index.ts
@@ -1,0 +1,4 @@
+export * from "./icon-name-helpers";
+export * from "./cell-background-helpers";
+export * from "./board-helpers";
+export * from "./storage-helpers";