+++ src/pages/game/interfaces/board-config.ts
@@ -1,0 +1,4 @@
+export interface BoardConfig {
+  boardSize: number;
+  numOfBombs: number;
+}