+++ src/pages/game/interfaces/config-info.ts
@@ -1,0 +1,4 @@
+export interface ConfigInfo {
+  boardSize: number;
+  numOfBombs: number;
+}