+++ src/pages/game/interfaces/index.ts
@@ -1,0 +1,3 @@
+export * from "./cell-item.interface";
+export * from "./config-info";
+export * from "./board-config";