+++ src/CHANGELOG.md
@@ -5,6 +5,38 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.6.0] - 2025-10-27
+
+### Changed
+
+- Create context to share common properties.
+
+## [0.5.0] - 2025-10-27
+
+### Added
+
+- Added functions to save and get storage from configuration.
+- Added Board, WinGame, and EndOfGame components.
+
+### Changed
+
+- change helpers file names.
+- Rename previous Board component to game.
+
+## [0.4.0] - 2025-10-27
+
+### Added
+
+- Added Mine interface and types.
+- Added cell component.
+- Added custom hook useCell.
+- Added game page and rename it to board page.
+
+### Changed
+
+- Updated README.md.
+- Rename Game references to Board in router.
+
 ## [0.3.0] - 2025-10-26
 
 ### Added